+++ NEWS.md
@@ -1,5 +1,15 @@
 # rABM (development version)
 
+# rABM 0.0.5
+2025-06-02
+* [new] modify_agents: A new function added.
+* [revised] 'init_agent' renamed to 'init_agents'
+* [revised] modify_G: argument 'field' renamed to 'new_obj'
+
+# rABM 0.0.4
+2025-05-31
+* [revised] modify_G: added a method 'add_agents'.
+
 # rABM 0.0.3
 2025-05-31
 * [revised] modify_G: put G and E to the formals when adding global_FUN, select_FUN, stop_FUN, or update_FUN.